+++ src/system/string.ts
@@ -36,8 +36,9 @@ export namespace Strings {
         return tokens;
     }
 
-    export function interpolate(template: string, context: object): string {
+    export function interpolate(template: string, context: object | undefined): string {
         if (!template) return template;
+        if (context === undefined) return template.replace(TokenSanitizeRegex, '');
 
         template = template.replace(TokenSanitizeRegex, '$${this.$1}');
         return new Function(`return \`${template}\`;`).call(context);